overhaul of a few functions

+++ src/ft_factorial.c
@@ -6,15 +6,15 @@
 /*   By: dromansk <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/02 17:08:46 by dromansk          #+#    #+#             */
-/*   Updated: 2018/12/02 17:09:59 by dromansk         ###   ########.fr       */
+/*   Updated: 2019/03/29 17:01:45 by dromansk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_factorial(int a)
+long long	ft_factorial(int a)
 {
-	int	b;
+	long long	b;
 
 	b = 1;
 	while (a)
